Don’t overwrite & re-fetch existing tiles

+++ src/mbtiles-db.js
@@ -2,12 +2,14 @@ const {fromCallback} = require('bluebird')
 const fs = require('fs')
 const MBTiles = require('@mapbox/mbtiles')
 
-async function createMBTiles(file) {
-  try {
-    await fs.promises.unlink(file)
-  } catch (ignored) {
-    // ignore error, likely file does not exist
+async function createMBTiles(file, overwriteIfPresent = false) {
+  if (overwriteIfPresent) {
+    try {
+      await fs.promises.unlink(file)
+    } catch (ignored) {
+    }
   }
+
   const mbtiles = await fromCallback(cb => new MBTiles(`${file}?mode=rwc`, cb))
 
   await fromCallback(cb => mbtiles.startWriting(cb))
@@ -28,12 +30,27 @@ async function createMBTiles(file) {
     return fromCallback(cb => mbtiles.putInfo(info, cb))
   }
 
+  async function hasTile(z, x, y) {
+    try {
+      await getTile(z, x, y)
+      return true
+    } catch (err) {
+      return false
+    }
+  }
+
+  async function getTile(z, x, y) {
+    return fromCallback(cb => mbtiles.getTile(z, x, y, cb))
+  }
+
   async function putTile(pbfTile, z, x, y) {
     return fromCallback(cb => mbtiles.putTile(z, x, y, pbfTile, cb))
   }
 
   return {
     putInfo,
+    hasTile,
+    getTile,
     putTile,
     close
   }
